Process only requested roster results (SMACK-538)
Prior to this change, Smack processes each RosterPacket (which is of
type IQ.Type.RESULT) as a roster result.

Any other client on the XMPP network can send such a packet (not
our server). This allows a malicious party to overwrite our Roster.

This patch changes smack so that a RosterPacket is discarded if it
not a reply to a roster request.